+++ src/pages/Mypage.tsx
@@ -1,4 +1,5 @@
 import styled from "styled-components";
+import MypageHonggu from "../components/mypage/MypageHonggu";
 
 const imgProfile = "images/profile.png";
 const Mypage = () => {
@@ -6,7 +7,6 @@ const Mypage = () => {
 
   return (
     <StyledDivContainer>
-      .
       <StyledDivMain>
         {/* MypageProfile 관련된 것 */}
         <StyledDivOne>
@@ -27,81 +27,9 @@ const Mypage = () => {
               </ProfileListReview>
             </ProfileListLikeReview>
           </ProfileList>
-          <div></div>
+          <ProfileEditbButton>수정하기</ProfileEditbButton>
         </StyledDivOne>
-        {/* MypageProfile 관련된 것 */}
-        <StyledDivTwo>
-          <button>관심</button>
-          <button>리뷰</button>
-        </StyledDivTwo>
-        <StyledDivThree>
-          <StyledDivReviewFoamMainContainer>
-            <StyledDivReviewFoamImgContainer>
-              <div></div>
-            </StyledDivReviewFoamImgContainer>
-            <StyledDivReviewFoamContentsContainer>
-              <div>
-                <p>⌥</p>
-              </div>
-              <div>
-                <p>닉네임</p>
-                <p>2023.01.01</p>
-              </div>
-              <div>
-                <p>
-                  리뷰내용 리뷰내용 리뷰내용리뷰내용 리뷰내용 리뷰내용리뷰내용
-                  리뷰내용 리뷰내용리뷰내용 리뷰내용 리뷰내용리뷰내용 리뷰내용
-                  리뷰내용리뷰내용 리뷰내용 리뷰내용리뷰내용 리뷰내용
-                  리뷰내용리뷰내용 리뷰내용 리뷰내용
-                </p>
-              </div>
-            </StyledDivReviewFoamContentsContainer>
-          </StyledDivReviewFoamMainContainer>
-          <StyledDivReviewFoamMainContainer>
-            <StyledDivReviewFoamImgContainer>
-              <div></div>
-            </StyledDivReviewFoamImgContainer>
-            <StyledDivReviewFoamContentsContainer>
-              <div>
-                <p>⌥</p>
-              </div>
-              <div>
-                <p>닉네임</p>
-                <p>2023.01.01</p>
-              </div>
-              <div>
-                <p>
-                  리뷰내용 리뷰내용 리뷰내용리뷰내용 리뷰내용 리뷰내용리뷰내용
-                  리뷰내용 리뷰내용리뷰내용 리뷰내용 리뷰내용리뷰내용 리뷰내용
-                  리뷰내용리뷰내용 리뷰내용 리뷰내용리뷰내용 리뷰내용
-                  리뷰내용리뷰내용 리뷰내용 리뷰내용
-                </p>
-              </div>
-            </StyledDivReviewFoamContentsContainer>
-          </StyledDivReviewFoamMainContainer>
-          <StyledDivReviewFoamMainContainer>
-            <StyledDivReviewFoamImgContainer>
-              <div></div>
-            </StyledDivReviewFoamImgContainer>
-            <StyledDivReviewFoamContentsContainer>
-              <div>
-                <p>⌥</p>
-              </div>
-              <div>
-                <p>닉네임</p>
-                <p>2023.01.01</p>
-              </div>
-              <div>
-                <p>
-                  리뷰내용 리뷰내용 리뷰내용리뷰내용 리뷰내용 리뷰내용리뷰내용
-                  리뷰내용 리뷰내용리뷰내용 리뷰내용 리뷰내용리뷰내용 리뷰내용
-                  리뷰내용리뷰내용 리뷰내용 리뷰내용리뷰내용 리뷰내용
-                  리뷰내용리뷰내용 리뷰내용 리뷰내용
-                </p>
-              </div>
-            </StyledDivReviewFoamContentsContainer>
-          </StyledDivReviewFoamMainContainer>
-        </StyledDivThree>
+        <MypageHonggu />
       </StyledDivMain>
     </StyledDivContainer>
   );
@@ -145,10 +73,9 @@ const ProfileList = styled.div`
 `;
 const ProfileNickname = styled.div`
   text-align: left;
-  height: 50px;
-  line-height: 50px;
   font-size: large;
   font-weight: 700;
+  padding: 5px;
 `;
 const ProfileListLikeReview = styled.div`
   display: grid;
@@ -163,106 +90,17 @@ const ProfileListReview = styled.div`
   font-size: small;
   text-align: center;
 `;
-// 버튼항목
-const StyledDivTwo = styled.div`
-  margin-top: 30px;
-  margin-bottom: 15px;
-
-  button:nth-child(1) {
-    margin-left: 10px;
-    width: 100px;
-    height: 40px;
-    font-size: 20px;
-    background: #e37b58;
-    border-radius: 30px;
-    padding: 10px 30px;
-    border: none;
-    &:hover {
-      background-color: #e37b58;
-      color: white;
-    }
-  }
-
-  button:nth-child(2) {
-    margin-left: 10px;
-    width: 100px;
-    height: 40px;
-    background: #ffffff;
-    border-radius: 30px;
-    padding: 10px 30px;
-    border: none;
-    font-size: 20px;
-    &:hover {
-      background-color: #e37b58;
-      color: white;
-    }
-  }
-`;
-
-// 리뷰창
-const StyledDivThree = styled.div`
-  padding: 20px;
-  border-radius: 30px;
-  background-color: #f5f5f5;
-  flex-direction: column;
-`;
-const StyledDivReviewFoamMainContainer = styled.div`
-  padding: 20px 0px 20px 0px;
-  background-color: white;
+const ProfileEditbButton = styled.button`
+  font-size: small;
+  text-align: center;
+  border: 0;
+  outline: none;
+  width: 80px;
+  height: 25px;
   border-radius: 10px;
-  display: flex;
-
-  margin: 0px 0px 20px 0px;
-`;
-
-const StyledDivReviewFoamImgContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 1em;
-  div:nth-child(1) {
-    width: 70px;
-    height: 70px;
-    border-radius: 70%;
-    overflow: hidden;
-    background-color: lightgray;
-  }
-`;
-
-const StyledDivReviewFoamContentsContainer = styled.div`
-  margin-right: 10px;
-
-  div:nth-child(1) {
-    display: flex;
-    justify-content: right;
-    align-items: center;
-    margin-right: 15px;
-    height: 30px;
-    p {
-      cursor: pointer;
-    }
-  }
-
-  div:nth-child(2) {
-    display: flex;
-
-    p:nth-child(1) {
-      font-weight: 700;
-      font-size: 1.2em;
-    }
-
-    p:nth-child(2) {
-      display: flex;
-      font-size: 0.8em;
-      justify-content: center;
-      align-items: center;
-      margin-left: 5px;
-      margin-bottom: 10px;
-    }
-  }
-
-  div:nth-child(3) {
-    display: flex;
+  :hover {
+    background-color: #e37b58;
+    transition: 0.7s;
   }
 `;
