+++ src/shared/Router.js
@@ -6,7 +6,6 @@ import Detail from "../pages/Detail";
 import Mypage from "../pages/Mypage";
 import Home from "../pages/Home";
 import Login from "../pages/Login";
-import MypageHonggu from "../pages/MypageHonggu";
 const Router = () => {
   return (
     <BrowserRouter>
@@ -16,7 +15,6 @@ const Router = () => {
         <Route path="/" element={<Main />} />
         <Route path="/:id" element={<Detail />} />
         <Route path="/mypage" element={<Mypage />} />
-        <Route path="/MypageHonggu" element={<MypageHonggu />} />
         <Route exact path="/login" element={<Login />} />
       </Routes>
     </BrowserRouter>
